refactor(access): move function get_parent_model to common viewset
ref: #982 #980
