+++ app/api/tests/unit/test_unit_common_viewset.py
@@ -430,6 +430,10 @@ class CommonViewSetTestCases(
                 'type': list,
                 'value': []
             },
+            'parent_model': {
+                'type': str,
+                'value': None
+            },
             'permission_classes': {
                 'type': list,
                 'value': [
